+++ src/classes/babies/ShopkeeperBaby.ts
@@ -1,7 +1,6 @@
 import {
   ModCallback,
   PickupPrice,
-  PickupVariant,
   RoomType,
 } from "isaac-typescript-definitions";
 import { Callback, inRoomType, levelHasRoomType } from "isaacscript-common";
@@ -13,14 +12,15 @@ export class ShopkeeperBaby extends Baby {
     return levelHasRoomType(RoomType.SHOP);
   }
 
-  @Callback(ModCallback.POST_PICKUP_INIT, PickupVariant.COLLECTIBLE)
-  postPickupInitCollectible(pickup: EntityPickup): void {
+  @Callback(ModCallback.POST_PICKUP_INIT)
+  postPickupInit(pickup: EntityPickup): void {
     if (
       pickup.Price !== (PickupPrice.NULL as int) &&
       pickup.Price !== (PickupPrice.FREE as int) &&
       inRoomType(RoomType.SHOP, RoomType.ERROR)
     ) {
       pickup.Price = PickupPrice.FREE;
+      pickup.AutoUpdatePrice = false;
     }
   }
 }