+++ docs/history.md
@@ -8,7 +8,7 @@
 - Changed the following babies:
   - Crow Baby (#12) - Soul of Eve effect on hit
   - Lowface Baby (#73) - Starts with Book of Virtues + Unicorn Stump
-  - Eyebrows Baby (#89) - Starts with 5x Guppy's Hair Ball + Blindfolded
+  - Eyebrows Baby (#89) - Starts with 5x Guppy's Hair Ball + flight + blindfolded
   - Lipstick Baby (#105) - Range oscillates
   - Tooth Baby (#478) - Starts with Black Tooth
   - Thirteenth Baby (#533) - Starts with Modeling Clay