+++ docs/history.md
@@ -12,6 +12,8 @@
   - Lowface Baby (#73) - Starts with Book of Virtues + Unicorn Stump
   - Eyebrows Baby (#89) - Starts with 5x Guppy's Hair Ball + flight + blindfolded
   - Lipstick Baby (#105) - Range oscillates
+  - Blisters Baby (#240) - Shot speed oscillates
+  - Cupcake Baby (#321) - Starts with Voodoo Head + Flat File
   - Burning Baby (#467) - Starts with Vasculitis + A Lighter + 40 Luck
   - Tooth Baby (#478) - Starts with Black Tooth
   - Thirteenth Baby (#533) - Starts with Modeling Clay