+++ docs/history.md
@@ -22,6 +22,7 @@
   - Thirteenth Baby (#533) - Starts with Modeling Clay
   - Painful Baby (#570) - Enemies respawn when reentering a room
   - Doleful Baby (#572) - Starts with Soy Milk + booger tears
+  - Cubic Baby (#573) - Starts with Brimstone Bombs + Scatter Bombs
   - Robo-Baby (#578) - Starts with Technology + 3x Robo-Baby
 - Thanks to Gamonymous for these new baby effects.
 - Fixed the bug where Spike Baby chests would not work properly. (Thanks SuperpaperChildy)